This is work to add support for the Linux Indexer.
As part of this I added the logic for the IndalekoLinuxMachineConfig.py so it now creates the config file and adds that to the database.

This also provides a prelminary version of the IndalekoLinuxLocalIndexer.py, though that remains a work in progress.

Some refactoring along the way as I pull some code into the generic layers.