More schema related work.
(1) I found a peculiar issue where I'd get an index constraint violation
on services, which didn't make much sense.  After digging, I found the
index creation service was a singleton, so it was creating exactly ONE
index.  It's no longer a singleton and indices are being createed.
(2) The ordering issue is that Services will construct its own entry.
This is not a good idea (never have the samee thing donee in two
different places.) The issue their was that it was using "name" as the
field for the index.  Since the data used "Name" the second service
entry would collide.  This is now resolved.
(3) The schema for the relationship data was just wrong.  I have revised
that code so it now generates a data modele driven schmea and the
reelationship data is now successfully being inserted into the database.

There is likely more work to do in cleaing up the schema stuff. The
actual creation of a new schema is much simpler and the data class code
generates data validators.

So, at least on Windows, I can once again index and store data in
ArangoDB for the local file system.